feat : update

+++ app/Http/Controllers/AssignPicController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\AssignPic;
+use App\Models\Project;
 use App\Models\User;
 use App\Traits\MessageResponse;
 use Illuminate\Http\Request;
@@ -13,12 +14,13 @@ use Illuminate\Support\Facades\Validator;
 class AssignPicController extends Controller
 {
     use MessageResponse;
-    protected $pic, $user;
+    protected $pic, $user, $project;
 
     public function __construct()
     {
         $this->pic = new AssignPic();
         $this->user = new User();
+        $this->project = new Project();
     }
 
     public function UserExist(Request $request)
@@ -34,6 +36,15 @@ class AssignPicController extends Controller
 
         try {
             DB::beginTransaction();
+
+            $exist = $this->pic
+                ->where('user_id', $request->user_id)
+                ->where('project_id', $request->project_id)->first();
+
+            if ($exist) {
+                $projectName = $this->project->where('id', $request->project_id)->value('name');
+                return $this->showFail("User already exists in project $projectName");
+            }
 
             $data = $this->pic->create([
                 'user_id'       => $request->user_id,
@@ -91,11 +102,28 @@ class AssignPicController extends Controller
         }
     }
 
-    public function destroy($id)
+    public function destroy(Request $request)
     {
+        $validate = Validator::make($request->all(), [
+            'user_id'    => 'integer|required',
+            'company_id' => 'integer|required'
+        ]);
+
+        if ($validate->fails()) {
+            return $this->showValidateError($validate->errors());
+        }
+
         try {
             DB::beginTransaction();
-            $pic = $this->pic->where('id', $id)->first();
+
+            $pic = $this->pic->whereHas('project', function ($query) use ($request) {
+                $query->where('company_id', $request->company_id);
+            })->where('user_id', $request->user_id)->first();
+
+            if (!$pic) {
+                return $this->showFail("User tidak ditemukan dalam perusahaan ini.");
+            }
+
             $pic->delete();
 
             DB::commit();
